organize Snap page objects
