+++ test/e2e/flask/user-operations.spec.ts
@@ -18,7 +18,7 @@ import { SWAP_TEST_ETH_USDC_TRADES_MOCK } from '../../data/mock-data';
 import { Mockttp } from '../mock-e2e';
 import TestDapp from '../page-objects/pages/test-dapp';
 import TestDappIndividualRequest from '../page-objects/pages/test-dapp-individual-request';
-import SnapAccountAbstractionKeyringPage from '../page-objects/pages/snap-account-abstraction-keyring-page';
+import SnapAccountAbstractionKeyringPage from '../page-objects/pages/snaps/account-abstraction-keyring-page';
 import ActivityTab from '../page-objects/pages/home/activity-tab';
 import TransactionConfirmation from '../page-objects/pages/confirmations/transaction-confirmation';
 import { mockSnapAccountAbstractionKeyRingAndSite } from '../mock-response-data/snaps/snap-local-sites/account-abstraction-keyring-site-mocks';
