+++ test/e2e/page-objects/pages/snaps/simple-keyring-page.ts
@@ -1,6 +1,6 @@
-import { Driver } from '../../webdriver/driver';
-import { WINDOW_TITLES } from '../../constants';
-import { getCleanAppState, regularDelayMs } from '../../helpers';
+import { Driver } from '../../../webdriver/driver';
+import { WINDOW_TITLES } from '../../../constants';
+import { getCleanAppState, regularDelayMs } from '../../../helpers';
 
 const SIMPLE_KEYRING_SNAP_ID = 'npm:@metamask/snap-simple-keyring-snap';
 
@@ -120,6 +120,9 @@ class SnapSimpleKeyringPage {
     tag: 'p',
   };
 
+  private readonly parentSelector =
+    '[data-testid="parent-selector-snap-simple-keyring-page"]';
+
   private readonly rejectRequestButton = {
     text: 'Reject Request',
     tag: 'button',
@@ -229,6 +232,7 @@ class SnapSimpleKeyringPage {
   async checkPageIsLoaded(): Promise<void> {
     try {
       await this.driver.waitForMultipleSelectors([
+        this.parentSelector,
         this.pageTitle,
         this.useSyncApprovalToggle,
       ]);
