+++ test/e2e/snaps/test-snap-homepage.spec.ts
@@ -4,7 +4,7 @@ import { DAPP_PATH, WINDOW_TITLES } from '../constants';
 import { withFixtures } from '../helpers';
 import FixtureBuilderV2 from '../fixtures/fixture-builder-v2';
 import HeaderNavbar from '../page-objects/pages/header-navbar';
-import SnapListPage from '../page-objects/pages/snap-list-page';
+import SnapListPage from '../page-objects/pages/snaps/list-page';
 import { login } from '../page-objects/flows/login.flow';
 import { openTestSnapClickButtonAndInstall } from '../page-objects/flows/install-test-snap.flow';
 import { mockHomePageSnap } from '../mock-response-data/snaps/snap-binary-mocks';
