+++ test/e2e/snaps/test-snap-management.spec.ts
@@ -1,7 +1,7 @@
 import { Driver } from '../webdriver/driver';
 import { TestSnaps } from '../page-objects/pages/test-snaps';
 import HomePage from '../page-objects/pages/home/homepage';
-import SnapListPage from '../page-objects/pages/snap-list-page';
+import SnapListPage from '../page-objects/pages/snaps/list-page';
 import FixtureBuilderV2 from '../fixtures/fixture-builder-v2';
 import { login } from '../page-objects/flows/login.flow';
 import { openTestSnapClickButtonAndInstall } from '../page-objects/flows/install-test-snap.flow';
