+++ test/e2e/snaps/test-snap-metrics.spec.ts
@@ -21,7 +21,7 @@ import { openTestSnapClickButtonAndInstall } from '../page-objects/flows/install
 import { TestSnaps } from '../page-objects/pages/test-snaps';
 import SnapInstall from '../page-objects/pages/dialog/snap-install';
 import HomePage from '../page-objects/pages/home/homepage';
-import SnapListPage from '../page-objects/pages/snap-list-page';
+import SnapListPage from '../page-objects/pages/snaps/list-page';
 import packageJson from '../../../package.json';
 
 async function mockedSnapInstall(mockServer: Mockttp) {
