+++ test/e2e/snaps/test-snap-update-component.spec.ts
@@ -7,7 +7,7 @@ import {
   mockWebpackPluginSnap,
 } from '../mock-response-data/snaps/snap-binary-mocks';
 import HeaderNavbar from '../page-objects/pages/header-navbar';
-import SnapListPage from '../page-objects/pages/snap-list-page';
+import SnapListPage from '../page-objects/pages/snaps/list-page';
 import { login } from '../page-objects/flows/login.flow';
 import { openTestSnapClickButtonAndInstall } from '../page-objects/flows/install-test-snap.flow';
 import { TestSnaps } from '../page-objects/pages/test-snaps';
