+++ test/e2e/tests/account/snap-account-contract-interaction.spec.ts
@@ -15,7 +15,7 @@ import { SMART_CONTRACTS } from '../../seeder/smart-contracts';
 import ActivityTab from '../../page-objects/pages/home/activity-tab';
 import HeaderNavbar from '../../page-objects/pages/header-navbar';
 import HomePage from '../../page-objects/pages/home/homepage';
-import SnapSimpleKeyringPage from '../../page-objects/pages/snap-simple-keyring-page';
+import SnapSimpleKeyringPage from '../../page-objects/pages/snaps/simple-keyring-page';
 import TestDapp from '../../page-objects/pages/test-dapp';
 import TransactionConfirmation from '../../page-objects/pages/confirmations/transaction-confirmation';
 import { installSnapSimpleKeyring } from '../../page-objects/flows/snap-simple-keyring.flow';
